Add files via upload
A Large Language Model (LLM) is an advanced AI algorithm that uses neural networks with extensive parameters for a variety of natural language processing tasks. Trained on large text datasets, LLMs excel in processing and generating human language, handling tasks such as text generation, translation, and summarization. Following is a review paper on the Large Language Model . Which reviews its Architecture and Application.